requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
certifi==2019.11.28
cffi==1.14.0
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
descartes==1.1.0
dominate==2.4.0
easydict==1.9
edict==0.51
entrypoints==0.3
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Dropzone==1.5.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mizani==0.6.0
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
olefile==0.46
opencv-python==3.4.2.17
palettable==3.3.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==6.2.2
plantcv==3.7
plotnine==0.5.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
scikit-image==0.14.2
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
spyder-kernels==0.5.2
statsmodels==0.11.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
traitlets==4.3.3
visitor==0.1.3
Vizer==0.1.5
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wincertstore==0.2
WTForms==2.2.1
zipp==3.0.0